Add Pvalue queries to the set of commands in SeekRPC (#27)
* Split SeekPValue.cpp into two files, one for main() initialization
* Allow to specify num random queries to us
* Add read/write 2D vector helper functions
* SeekPValue Load from vector file
* Add pvalue to the SeekRPC interface
* SeekRPC add pvalue initialization, increment RPCversion
* PValue add useRank and checks for num genes and num ranks or scores
* Random query generator for pvalue server
* Pvalue ones-based ranking
* Add test inputs for pvalue tests
* Add SeekRPC pvalue tests for score and rank based pvalue queries
* Add other pvalue unit tests
* add test for read-write 2D vector

+++ tools/SeekRPC/SeekRPCClient.cpp
@@ -149,7 +149,7 @@ int main(int argc, char **argv)
         seekClient.ping();
 
         // Check for version compatibility
-        int32_t version = seekClient.getRpcVersion();
+        double version = seekClient.getRpcVersion();
         assert(version == g_seek_rpc_constants.RPCVersion);
 
         if (!args.outputFile.empty()) {
